refactor(desktop): isolate renderable file protocol
